Doc-testing (#233)
bug fixes, doc updates
