+++ CHANGELOG.md
@@ -5,8 +5,9 @@
 ![](https://img.shields.io/badge/release%20date-April%2013%2C%202025-blue)
 ![GitHub release](https://img.shields.io/github/downloads-pre/dovrosenberg/fvtt-campaign-builder/v0.1.2/module.zip)
 
-- Bug fix - collapse all now works properly
 - Added buttons to add actors/scenes and drag indicator to scenes in Entry document relationship tabs
+- Lots of documentation improvements
+- Various minor bug fixes
 
 
 ## v0.1.1 - Tags and quality of life
