+++ src/applications/stores/topicDirectoryStore.ts
@@ -555,7 +555,6 @@ export const useTopicDirectoryStore = defineStore('topicDirectory', () => {
   // this includes: all nodes matching the filterText, all of their ancestors, and
   //    all of their types (we also ways leave the packs)
   // it's an object keyed by topic with a list of all the ids to include
-  // TODO - a checkbox option that uses search to filter by all searchable fields vs just name
   const updateFilterNodes = (): void => {
     if (!currentWorld.value)
       return;
@@ -578,6 +577,7 @@ export const useTopicDirectoryStore = defineStore('topicDirectory', () => {
       // filter on name and type
       let matchedEntries = topicObj.filterEntries((e: Entry)=>( filterText.value === '' || regex.test( e.name || '' ) || regex.test( e.type || '' )))
         .map((e: Entry): string=>e.uuid) as string[];
+
   
       // add the ancestors and types; iterate backwards so that we can push on the end and not recheck the ones we're adding
       for (let j=matchedEntries.length-1; j>=0; j--) {
