+++ src/components/Tags.vue
@@ -184,21 +184,26 @@
         whitelist.push(tag);
     }
 
-    var input = document.getElementById("fcb-tags-input") as HTMLInputElement;
-    tagify.value = new Tagify(input, {
-      whitelist: getWhitelist(),
-      dropdown: {
-        enabled: 1,
-        position: 'text',
-        searchKeys: ['value'],
-        tabKey: true,
-      },
-      transformTag: transformTag,
-      callbacks: {
-        add: (e) => { onTagAdded(e); },
-        remove: (e) => { onTagRemoved(e); },
-      }
-    })
+    // Use setTimeout to ensure the DOM is fully rendered
+    setTimeout(() => {
+      var input = document.getElementById("fcb-tags-input") as HTMLInputElement;
+
+      tagify.value = new Tagify(input, {
+        whitelist: getWhitelist(),
+        dropdown: {
+          enabled: 1,
+          position: 'text',
+          searchKeys: ['value'],
+          tabKey: true,
+        },
+        transformTag: transformTag,
+        callbacks: {
+          add: (e) => { onTagAdded(e); },
+          remove: (e) => { onTagRemoved(e); },
+        }
+      });
+    }, 100);
+
   });
 </script>
