+++ src/hooks/index.ts
@@ -1,11 +1,11 @@
 import { registerFori18nInitHook } from './i18Init';
 import { registerForInitHook } from './init';
 import { registerForReadyHook } from './ready';
-import { registerForUpdateActorHook } from './updateActor';
+import { registerForUpdateHooks } from './updateDocuments';
 
 export function registerForHooks() {
   registerForInitHook();
   registerForReadyHook();
   registerFori18nInitHook();
-  registerForUpdateActorHook();
+  registerForUpdateHooks();
 }
